+++ src/store/index.ts
@@ -1,2 +1,3 @@
-export * from './store';
-export * from './types'; 
+export { store, persistor } from './store';
+export type { RootState } from './types';
+export { useAppDispatch, useAppSelector } from './hooks';